Make sure that encodeResource always sets the type of the passed Resource
Previously, it was required to always pass the type in the values, which
is not how patch is usually used.

This will now set the type of the Resource instead of from the values.

It still keeps the validation in place to check that IF you pass a type
in the values, it errors if it doesn't match with the Resource type.